# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Book": {
        "properties": {
          "author": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "publication_date": {
            "format": "date",
            "type": "string"
          },
          "publication_year": {
            "type": "integer"
          },
          "summary": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "author",
          "publication_date",
          "publication_year",
          "summary",
          "created_at"
        ],
        "type": "object"
      },
...
      "CreateBookRequest": {
        "properties": {
          "author": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "publication_date": {
            "format": "date",
            "type": "string"
          },
          "publication_year": {
            "type": "integer"
          },
          "summary": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "author",
          "summary"
        ],
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "type": "object"
          },
          "request_id": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "error",
          "request_id"
        ],
        "type": "object"
      },
...
      "ImportBookResponse": {
        "properties": {
          "book": {
            "$ref": "#/components/schemas/Book"
          },
          "extracted": {
            "properties": {
              "author": {
                "type": "string"
              },
              "category": {
                "type": "string"
              },
              "publication_year": {
                "type": "integer"
              },
              "summary": {
                "type": "string"
              },
              "title": {
                "type": "string"
              }
            },
            "required": [
              "title",
              "category",
              "author",
              "publication_year",
              "summary"
            ],
            "type": "object"
          }
        },
        "required": [
          "book",
          "extracted"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Shared API_KEY configured by the operator. Optional only when the backend runs locally with API_KEY unset.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Flask API for the backend AI challenge: virtual library, Python assistant, LangSmith feedback, uploads and semantic search.",
    "title": "MindSight AI API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/books": {
      "get": {
        "parameters": [
          {
            "description": "Filter by title",
            "in": "query",
            "name": "title",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by author",
            "in": "query",
            "name": "author",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by category",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Search title, author, category and summary",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Max items to return (1-200). Omit for the full unbounded list.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Items to skip before collecting results. Omit to start at 0.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  },
                  "type": "array"
                }
              }
            },
            "description": "JSON array response"
          }
        },
        "summary": "Search books",
        "tags": [
          "Books"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            },
            "description": "JSON response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON response"
          }
        },
        "summary": "Create a book",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/import": {
      "post": {
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "description": ".txt, .md, .json or .pdf containing title, author, category, year and summary. Content and declared MIME are validated; PDFs are subject to configured resource limits.",
                    "format": "binary",
                    "type": "string"
                  }
                },
                "required": [
                  "file"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBookResponse"
                }
              }
            },
            "description": "JSON response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON response"
          }
        },
        "summary": "Import a book from uploaded file",
        "tags": [
          "Books"
        ]
      }
    },
    "/api/v1/books/{book_id}": {
      "get": {
        "parameters": [
          {
            "description": "Book id",
            "in": "path",
            "name": "book_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            },
            "description": "JSON response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "JSON response"
          }
        },
        "summary": "Get a book by id",
        "tags": [
          "Books"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Local Flask",
      "url": "http://localhost:5000"
    },
    {
      "description": "Vite proxy / Docker frontend",
      "url": "http://localhost:3002"
    }
  ],
...
}